drivers: mediatek: ft5626: enable multi touch protocol
The driver already supports the multi touch protocol B but this was
not enabled. Also it were missing the slot initialization which was
hidden behind a kernel version check which didn't applied anymore
cause of a changed API.

With all this the multi touch events are now properly recognized by
libinput and can be processed.

Signed-off-by: Simon Fels <simon.fels@canonical.com>