binder: search for receiving context pid in global namespace
BugLink: https://bugs.launchpad.net/avila-private/+bug/1546179

If binder cannot find the receiving task in the sender's namespace it
will default to a null PID which thus breaks communication in our
container-based scenario. This patch extends binder to first look in the
senders namespace and if unsuccesfull search in the global namespace.