build: add missing toolchain files
The toolchain that we added to this repository is incomplete; it
probably worked because some files were being picked up from the host
machine, but that's not working anymore (at least when building on a
20.04 machine).

I'm adding here the files that are needed to complete a successful
build. They are taken from the [android
prebuilts](https://android.googlesource.com/platform/prebuilts/gcc/linux-x86/aarch64/aarch64-linux-android-4.9/),
and specifically from the `android-6.0.1_r81` tag (many other tags would
probably work as well, but I preferred to pick one from the same epoch
as the M10).

Note that there are more files in that prebuilts archive that we are
missing here. If we meet other build issues in the future, we might want
to add all of them.